Remove server TLS configuration (#40)
It is no longer required to make use of the instance identity
certificates on cloud.gov. Instead, apps are expected to take advantage
of the Envoy proxies, which make applications available over the
application_route:61443. This port terminates TLS before reaching the
application, so no cert config is required on the application.